requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
Flask==1.1.2
Flask-WTF==0.14.3
future==0.18.2
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.3
opencv-python==4.4.0.46
pathlib2==2.3.5
Pillow==8.0.1
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.11.26
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
